requirements.txt
dspy-ai==2.4.10
certifi==2021.5.30
chardet==4.0.0
click==8.1.7
colorama==0.4.6
docx2txt==0.8
idna==2.10
lxml==5.2.2
mypy-extensions==1.0.0
oauthlib==3.1.1
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
python-docx==1.1.2
requests==2.25.1
requests-oauthlib==1.3.0
tomli==2.0.1
twitter==1.19.2
camelot-py==0.11.0
python-dotenv==1.0.1
opencv-python==4.10.0.84
ghostscript==0.7
